I split off functionality from both the color sensor and microwave handler and gave that to the ColourSensorInterpreter. Also making sure the MicrowaveHandler can properly interpret what the ColourSensorInterpreter is saying.
